Implemented exception handling middleware

+++ src/EventsWepApp/EventsWebApp.Application/Commands/Event/DeleteEventCommand/DeleteEventCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using EventsWebApp.Domain.Interfaces;
+﻿using EventsWebApp.Application.Exceptions;
+using EventsWebApp.Domain.Interfaces;
 using FluentValidation;
 using MediatR;
 using System;
@@ -27,6 +28,12 @@ namespace EventsWebApp.Application.Commands.Event.DeleteEventCommand
             if (!validationResult.IsValid)
             {
                 throw new ValidationException(validationResult.Errors);
+            }
+
+            var existingEvent = await _eventRepository.GetByIdAsync(request.id, cancellationToken: cancellationToken);
+            if (existingEvent == null)
+            {
+                throw new NotFoundException("Event with given Id does not exist.");
             }
 
             await _eventRepository.DeleteAsync(request.id, cancellationToken);
